requirements.txt
fastapi
uvicorn[standard]
python-dotenv
websockets
Jinja2

langchain
langchain-openai
langchain-community
langsmith
requests
httpx



